# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WireGuard Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nodes:
    get:
      operationId: listNodes
      tags: [nodes]
      summary: List all VPN nodes
      responses:
        "200":
          description: List of nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
    post:
      operationId: createNode
      tags: [nodes]
      summary: Add a new VPN exit node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeInput"
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
  /nodes/{nodeId}:
    get:
      operationId: getNode
      tags: [nodes]
      summary: Get a single node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Node detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          description: Not found
    patch:
      operationId: updateNodeSsh
      tags: [nodes]
      summary: Update SSH credentials for a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeSshUpdateInput"
      responses:
        "200":
          description: Updated node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          description: Not found
    delete:
      operationId: deleteNode
      tags: [nodes]
      summary: Remove a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /nodes/{nodeId}/test-connection:
    post:
      operationId: testNodeConnection
      tags: [nodes]
      summary: Test SSH connection to a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"
  /nodes/{nodeId}/setup-wireguard:
    post:
      operationId: setupWireguard
      tags: [nodes]
      summary: Install and configure WireGuard on a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WireguardSetupInput"
      responses:
        "200":
          description: Setup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /nodes/{nodeId}/import-wireguard:
    post:
      operationId: importWireguard
      tags: [nodes]
      summary: Import existing WireGuard configuration from a node without reinstalling
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /nodes/{nodeId}/wireguard/start:
    post:
      operationId: startWireguard
      tags: [nodes]
      summary: Start WireGuard on a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /nodes/{nodeId}/wireguard/stop:
    post:
      operationId: stopWireguard
      tags: [nodes]
      summary: Stop WireGuard on a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /nodes/{nodeId}/wireguard/restart:
    post:
      operationId: restartWireguard
      tags: [nodes]
      summary: Restart WireGuard on a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /nodes/{nodeId}/wireguard/status:
    post:
      operationId: statusWireguard
      tags: [nodes]
      summary: Get live systemd status of WireGuard on a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /nodes/{nodeId}/wireguard:
    delete:
      operationId: uninstallWireguard
      tags: [nodes]
      summary: Remove WireGuard from a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
...
components:
  schemas:
...
    Node:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        ipAddress:
          type: string
        sshUser:
          type: string
        sshPort:
          type: integer
        sshAuthMethod:
          type: string
          enum: [password, private_key]
        wireguardPort:
          type: integer
        wireguardStatus:
          type: string
          enum: [not_installed, stopped, running, unknown]
        wireguardInterface:
          type: string
        clientCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - ipAddress
        - sshUser
        - sshPort
        - sshAuthMethod
        - wireguardPort
        - wireguardStatus
        - wireguardInterface
        - clientCount
        - createdAt
    NodeInput:
      type: object
      properties:
        name:
          type: string
        ipAddress:
          type: string
        sshUser:
          type: string
        sshPort:
          type: integer
        sshAuthMethod:
          type: string
          enum: [password, private_key]
        sshPassword:
          type: string
        sshPrivateKey:
          type: string
      required:
        - name
        - ipAddress
        - sshUser
    NodeSshUpdateInput:
      type: object
      properties:
        sshAuthMethod:
          type: string
          enum: [password, private_key]
        sshUser:
          type: string
        sshPort:
          type: integer
        sshPassword:
          type: string
        sshPrivateKey:
          type: string
    WireguardSetupInput:
      type: object
      properties:
        port:
          type: integer
        interface:
          type: string
        subnet:
          type: string
      required:
        - port
    ConnectionTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        latencyMs:
          type: ["integer", "null"]
      required:
        - success
        - message
    ImportResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        wireguardInterface:
          type: ["string", "null"]
        wireguardPort:
          type: ["integer", "null"]
        wireguardSubnet:
          type: ["string", "null"]
        wireguardStatus:
          type: ["string", "null"]
        existingPeerCount:
          type: ["integer", "null"]
        importedClientCount:
          type: ["integer", "null"]
    ActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        output:
          type: ["string", "null"]
      required:
        - success
        - message
...